requirements.txt
Pillow
easydev
colormap
